requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
ale-py==0.8.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==23.1.0
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
black==23.7.0
blessed==1.20.0
box2d-py==2.3.5
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
cloudpickle==2.2.1
cmake==3.26.4
colorful==0.5.5
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.7
dm-tree==0.1.8
exceptiongroup==1.1.3
Farama-Notifications==0.0.4
filelock==3.12.2
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.10.0
glfw==2.6.2
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gpustat==1.1
grpcio==1.51.3
gymnasium==0.29.0
h5py==3.9.0
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.18.4
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
line-profiler==4.1.1
lit==16.0.6
llvmlite==0.43.0
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
msgpack==1.0.5
mujoco==2.3.7
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.60.0
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.77
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencensus==0.11.2
opencensus-context==0.1.3
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.9.1
plotly==5.18.0
pluggy==1.3.0
prometheus-client==0.17.1
protobuf==4.23.4
psutil==5.9.5
py-spy==0.3.14
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.0
pydantic_core==2.3.0
pygame==2.5.0
Pygments==2.16.1
PyOpenGL==3.1.7
pyparsing==3.0.9
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
ray==2.5.1
referencing==0.29.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.5.1
scipy==1.11.3
Shimmy==1.1.0
six==1.16.0
smart-open==6.3.0
stable-baselines3==2.1.0
swig==4.1.1
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensordict==0.2.1
threadpoolctl==3.5.0
tomli==2.0.1
torch==2.1.0
torchrl==0.2.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.1.0
tslearn==0.6.3
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.21.0
wcwidth==0.2.6
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
